# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEVIUM AI Writing Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for the authenticated user
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects, dashboard]
      summary: Get project summary stats (books, chapters, scenes, word count)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
...
  /projects/{projectId}/locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List locations in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a location
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /projects/{projectId}/locations/{id}:
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationBody"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── World Rules ───────────────────────────────────────────────────────────
...
  /projects/{projectId}/memory:
    get:
      operationId: listMemoryItems
      tags: [memoryItems]
      summary: List memory items in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of memory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryItem"
    post:
      operationId: createMemoryItem
      tags: [memoryItems]
      summary: Create a memory item
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryItemBody"
      responses:
        "201":
          description: Created memory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryItem"
  /projects/{projectId}/memory/{id}:
    patch:
      operationId: updateMemoryItem
      tags: [memoryItems]
      summary: Update a memory item
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemoryItemBody"
      responses:
        "200":
          description: Updated memory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryItem"
    delete:
      operationId: deleteMemoryItem
      tags: [memoryItems]
      summary: Delete a memory item
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Timeline Events ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [novel, saga, articles, screenplay, other]
        primaryLanguage:
          type: string
        status:
          type: string
          enum: [active, archived, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - type
        - primaryLanguage
        - status
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [novel, saga, articles, screenplay, other]
        primaryLanguage:
          type: string
        status:
          type: string
          enum: [active, archived, completed]
      required:
        - name
        - type
        - primaryLanguage
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [novel, saga, articles, screenplay, other]
        primaryLanguage:
          type: string
        status:
          type: string
          enum: [active, archived, completed]
    ProjectSummary:
      type: object
      properties:
        projectId:
          type: integer
        totalBooks:
          type: integer
        totalChapters:
          type: integer
        totalScenes:
          type: integer
        totalWords:
          type: integer
        totalCharacters:
          type: integer
        lastEditedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - projectId
        - totalBooks
        - totalChapters
        - totalScenes
        - totalWords
        - totalCharacters
        - lastEditedAt
...
    Location:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        significance:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - createdAt
        - updatedAt
    CreateLocationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        significance:
          type: ["string", "null"]
      required:
        - name
    UpdateLocationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        significance:
          type: ["string", "null"]

    # ── WorldRule ─────────────────────────────────────────────────────────────
...
    MemoryItem:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
          enum: [event, injury, secret, relationship, death, promise, mystery, location_change, knowledge, other]
        title:
          type: string
        content:
          type: string
        scope:
          type: string
          enum: [global, book, chapter, scene]
        status:
          type: string
          enum: [suggested, canonical, discarded]
        confidence:
          type: ["integer", "null"]
        sourceSceneId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - title
        - content
        - scope
        - status
        - createdAt
        - updatedAt
    CreateMemoryItemBody:
      type: object
      properties:
        type:
          type: string
          enum: [event, injury, secret, relationship, death, promise, mystery, location_change, knowledge, other]
        title:
          type: string
        content:
          type: string
        scope:
          type: string
          enum: [global, book, chapter, scene]
        status:
          type: string
          enum: [suggested, canonical, discarded]
        confidence:
          type: ["integer", "null"]
        sourceSceneId:
          type: ["integer", "null"]
      required:
        - type
        - title
        - content
        - scope
    UpdateMemoryItemBody:
      type: object
      properties:
        type:
          type: string
          enum: [event, injury, secret, relationship, death, promise, mystery, location_change, knowledge, other]
        title:
          type: string
        content:
          type: string
        scope:
          type: string
          enum: [global, book, chapter, scene]
        status:
          type: string
          enum: [suggested, canonical, discarded]
        confidence:
          type: ["integer", "null"]

    # ── TimelineEvent ─────────────────────────────────────────────────────────
...